Refactor download managers
